Port new ducks to stock redux format
