+++ src/store/reducers/app.js
@@ -1,17 +1,10 @@
-// Constants
-const ROUTE_MODE = 'ROUTE'
-const REGION_MODE = 'REGION'
+import { SET_ANALYSIS_MODE, CLEAR_ANALYSIS_MODE } from '../actions'
 
-// Actions
-const SET_ANALYSIS_MODE = 'analyst-ui/app/SET_ANALYSIS_MODE'
-const CLEAR_ANALYSIS_MODE = 'analyst-ui/app/CLEAR_ANALYSIS_MODE'
-
-// Reducer
 const initialState = {
   analysisMode: null
 }
 
-export default function reducer (state = initialState, action) {
+const app = (state = initialState, action) => {
   switch (action.type) {
     case SET_ANALYSIS_MODE:
       return {
@@ -28,22 +21,4 @@ export default function reducer (state = initialState, action) {
   }
 }
 
-// Action creators
-function setAnalysisMode (mode) {
-  return {
-    type: SET_ANALYSIS_MODE,
-    mode
-  }
-}
-
-export function setRegionAnalysisMode () {
-  return setAnalysisMode(REGION_MODE)
-}
-
-export function setRouteAnalysisMode () {
-  return setAnalysisMode(ROUTE_MODE)
-}
-
-export function clearAnalysisMode () {
-  return { type: CLEAR_ANALYSIS_MODE }
-}
+export default app
