+++ src/store/reducers/viewBounds.js
@@ -1,20 +1,17 @@
-// Actions
-const SET_BOUNDS = 'analyst-ui/viewBounds/SET_BOUNDS'
-const CLEAR_BOUNDS = 'analyst-ui/viewBounds/CLEAR_BOUNDS'
+import { SET_VIEW_BOUNDS, CLEAR_VIEW_BOUNDS } from '../actions'
 
-// Reducer
 const initialState = {
   bounds: null
 }
 
-export default function reducer (state = initialState, action) {
+const viewBounds = (state = initialState, action) => {
   switch (action.type) {
-    case SET_BOUNDS:
+    case SET_VIEW_BOUNDS:
       return {
         ...state,
         bounds: action.bounds
       }
-    case CLEAR_BOUNDS:
+    case CLEAR_VIEW_BOUNDS:
       return {
         ...state,
         bounds: null
@@ -24,14 +21,4 @@ export default function reducer (state = initialState, action) {
   }
 }
 
-// Action creators
-export function setBounds (bounds) {
-  return {
-    type: SET_BOUNDS,
-    bounds
-  }
-}
-
-export function clearBounds () {
-  return { type: CLEAR_BOUNDS }
-}
+export default viewBounds
